--- bundle/src/main/java/com/adobe/acs/commons/indesign/dynamicdeckdynamo/servlets/package-info.java
@@ -1,23 +1,0 @@
-/*
- * #%L
- * ACS AEM Commons Bundle
- * %%
- * Copyright (C) 2016 Adobe
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-@Version("1.0.0")
-package com.adobe.acs.commons.indesign.dynamicdeckdynamo.servlets;
-
-import org.osgi.annotation.versioning.Version;
